update to version 0.4.0
Allow the user to input key commands without blocking the main animation
loop. The key bindings allow the user to control and alter the runtime.

Add the ability to define headers in the asciimation file.

Use the headers to check if the screen is the right size to run the
animation.

Update the examples to use the new file format

Update the readme with more information on usage.